backend/services/image_processor.py
import os
import uuid
import base64
from PIL import Image
import io
import anthropic
from typing import List, Dict, Any

class ImageProcessor:
    """Service for processing images through Omniparser and Claude"""
    
    def __init__(self, anthropic_api_key: str = None):
        self.anthropic_client = None
        if anthropic_api_key:
            self.anthropic_client = anthropic.Anthropic(api_key=anthropic_api_key)
    
    def process_image_with_omniparser(self, image_path: str) -> Dict[str, Any]:
        """
        Process image through Microsoft Omniparser
        TODO: Implement actual Omniparser integration
        """
        # Placeholder implementation
        return {
            'success': True,
            'parsed_data': {
                'objects': ['tree', 'sky', 'grass'],
                'text': [],
                'metadata': {
                    'confidence': 0.85,
                    'processing_time': '2.3s'
                }
            }
        }
    
    def generate_labels_with_claude(self, image_path: str, omniparser_data: Dict = None) -> List[str]:
        """
        Generate image labels using Claude AI
        """
        if not self.anthropic_client:
            # Return mock labels if no API key configured
            return ['photo', 'image', 'visual', 'content']
        
        try:
            # Convert image to base64 for Claude
            with open(image_path, 'rb') as image_file:
                image_data = base64.b64encode(image_file.read()).decode('utf-8')
            
            # Prepare prompt for Claude
            prompt = """
            Analyze this image and provide relevant tags/labels that describe its content. 
            Focus on:
            - Objects and subjects in the image
            - Setting/environment (indoor, outdoor, etc.)
            - Activities or actions
            - Mood or atmosphere
            - Colors and visual elements
            
            Return only a comma-separated list of relevant tags, no explanations.
            """
            
            if omniparser_data:
                prompt += f"\n\nAdditional context from image parsing: {omniparser_data}"
            
            # TODO: Implement actual Claude API call with image
            # For now, return mock labels
            mock_labels = [
                'outdoor', 'nature', 'landscape', 'photography', 
                'scenic', 'natural', 'environment', 'beautiful'
            ]
            
            return mock_labels
            
        except Exception as e:
            print(f"Error generating labels with Claude: {e}")
            return ['photo', 'image', 'unprocessed']
    
    def generate_detailed_description(self, image_path: str) -> str:
        """
        Generate detailed description of image using Claude
        """
        if not self.anthropic_client:
            return "Detailed image analysis requires Claude API configuration."
        
        try:
            # TODO: Implement actual Claude API call for detailed description
            return """
            This image shows a beautiful natural landscape with excellent composition and lighting. 
            The photograph captures the essence of outdoor beauty with rich colors and clear details. 
            The scene appears to be taken during optimal lighting conditions, creating a visually 
            appealing and well-balanced composition.
            """
            
        except Exception as e:
            print(f"Error generating description with Claude: {e}")
            return "Unable to generate detailed description at this time."
    
    def create_thumbnail(self, image_path: str, thumbnail_size: tuple = (200, 200)) -> str:
        """
        Create thumbnail for uploaded image
        """
        try:
            with Image.open(image_path) as img:
                # Create thumbnail
                img.thumbnail(thumbnail_size, Image.Resampling.LANCZOS)
                
                # Generate thumbnail filename
                base_name = os.path.splitext(os.path.basename(image_path))[0]
                thumbnail_path = os.path.join(
                    os.path.dirname(image_path), 
                    f"{base_name}_thumb.jpg"
                )
                
                # Save thumbnail
                img.save(thumbnail_path, "JPEG", quality=85)
                return thumbnail_path
                
        except Exception as e:
            print(f"Error creating thumbnail: {e}")
            return None
