backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
boto3==1.28.85
anthropic==0.7.8
Pillow==10.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
